+++ src/main/java/frc/robot/subsystems/coraller/Coraller.java
@@ -6,10 +6,6 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
-import frc.robot.subsystems.coraller.CorallerSetpoints.AnglerSetpoints;
-import frc.robot.subsystems.coraller.CorallerSetpoints.ElevatorSetpoints;
-import frc.robot.subsystems.coraller.CorallerSetpoints.FlickerSetpoints;
-
 public class Coraller extends SubsystemBase {
   private final Elevator elevator = new Elevator();
   private final Angler angler = new Angler();
@@ -129,30 +125,5 @@ public class Coraller extends SubsystemBase {
     builder.setSmartDashboardType(getName());
     builder.setSafeState(this::stop);
     builder.setActuator(true);
-  }
-
-  /** Heights and angles for the Coraller's possible states. */
-  private enum CorallerState {
-    STOW(ElevatorSetpoints.STOW_HEIGHT, AnglerSetpoints.STOW_ANGLE, FlickerSetpoints.STOW_ANGLE),
-    L1(ElevatorSetpoints.L1_HEIGHT, AnglerSetpoints.L1_ANGLE, FlickerSetpoints.STOW_ANGLE),
-    L2(ElevatorSetpoints.L2_HEIGHT, AnglerSetpoints.L2_ANGLE, FlickerSetpoints.STOW_ANGLE),
-    L2_ALGAE(ElevatorSetpoints.L2_HEIGHT, AnglerSetpoints.L2_ANGLE, FlickerSetpoints.L2_ANGLE),
-    L3(ElevatorSetpoints.L3_HEIGHT, AnglerSetpoints.L3_ANGLE, FlickerSetpoints.STOW_ANGLE),
-    L3_ALGAE(ElevatorSetpoints.L3_HEIGHT, AnglerSetpoints.L3_ANGLE, FlickerSetpoints.L3_ANGLE),
-    L4(ElevatorSetpoints.L4_HEIGHT, AnglerSetpoints.L4_ANGLE, FlickerSetpoints.STOW_ANGLE),
-    RECEIVE(ElevatorSetpoints.RECEIVE_HEIGHT, AnglerSetpoints.RECEIVE_ANGLE, FlickerSetpoints.STOW_ANGLE);
-
-    /** Inches */
-    public final double elevatorPosition;
-    /** Degrees */
-    public final double anglerPosition;
-    /** Degrees */
-    public final double flickerPosition;
-
-    private CorallerState(double elevatorPosition, double anglerPosition, double flickerPosition) {
-      this.elevatorPosition = elevatorPosition;
-      this.anglerPosition = anglerPosition;
-      this.flickerPosition = flickerPosition;
-    }
   }
 }
